Renamed wiz_v2 to wiz. Removed old connector implementation.
