+++ tasks/connectors/wiz/lib/vulns_mapper.rb
@@ -3,7 +3,7 @@
 require_relative "mapper"
 module Kenna
   module Toolkit
-    module WizV2
+    module Wiz
       class VulnsMapper < Mapper
         def name
           "Vuln"
