+++ tasks/connectors/wiz/readme.md
@@ -1,11 +1,12 @@
-## Running the Wiz task 
+## Running the WIZ task 
 
-This toolkit brings in data from Wiz
+This toolkit brings in data from WIZ for Issues and Vulnerabilities
 
-To run this task you need the following information from Wiz: 
+To run this task you need the following information from WIZ: 
 
-1. Client Id
-1. Client Secret
+1. WIZ client id
+2. WIZ client secret
+3. WIZ API hostname
 
 ## Command Line
 
@@ -13,25 +14,31 @@ See the main Toolkit for instructions on running tasks. For this task, if you le
 
 Recommended Steps: 
 
-1. Run with Wiz Keys only to ensure you are able to get data properly from the scanner
+1. Run with WIZ Keys only to ensure you are able to get data properly from the scanner
 1. Review output for expected data
-1. Create Kenna Data Importer connector in Kenna (example name: Wiz KDI) 
+1. Create Kenna Data Importer connector in Kenna (example name: WIZ KDI) 
 1, Manually run the connector with the json from step 1 
 1. Click on the name of the connector to get the connector id
-1. Run the task with Wiz Keys and Kenna Key/connector id
+1. Run the task with WIZ Keys and Kenna Key/connector id
 
 
 
-Complete list of Options:
+Complete  list  of Opt ions :
 
-| Option | Required | Description | default |
-| --- | --- | --- | --- |
-| wiz_client_id | true | Wiz Client ID | n/a |
-| wiz_client_secret | true | Wiz Client Secret | n/a |
-| wiz_api_host | false | url to generate and retrieve Wiz reports | u3.test |
-| wiz_auth_endpoint | false | url for Wiz authentication | auth0.test |
-| report_object_types | false | list of desired report types | VIRTUAL_MACHINE,CONTAINER_IMAGE,SERVERLESS |
-| kenna_api_key | false | Kenna API Key for use with connector option | n/a |
-| kenna_api_host | false | Kenna API Hostname if not US shared | api.kennasecurity.com |
-| kenna_connector_id | false | If set, we'll try to upload to this connector | n/a |
-| output_directory | false | If set, will write a file upon completion. Path is relative to #{$basedir} | output/wiz |
+| Option             | Required | Description                                                                                                                                                                                                                                                | default            |
+|--------------------|----------|------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|--------------------|
+| wiz_client_id      | true     | WIZ client id                                                                                                                                                                                                                                              | n/a                |
+| wiz_client_secret  | true     | WIZ client secret                                                                                                                                                                                                                                          | n/a                |
+| wiz_api_host       | true     | WIZ API Endpoint URL. If schema is included, it should be between double quotes escaped.                                                                                                                                                                   | n/a                |
+| wiz_auth_endpoint  | false    | WIZ auth endpoint hostname used to get the authorization token, e.g. auth.wiz.io or auth0.test.wiz.io                                                                                                                                                      | auth.wiz.io        |
+| wiz_page_size      | false    | Maximum number of issues to retrieve in each page. Max value is 500.                                                                                                                                                                                       | 500                |
+| days_back          | false    | Integer days number to get the vulnerabilities/issues detected x days back TODAY. Import all history if not present                                                                                                                                        | n/a                |
+| vuln_object_types  | false    | Array of object types for VULNS import. Allowed values: VIRTUAL_MACHINE,CONTAINER_IMAGE,SERVERLESS. Import all if not present.                                                                                                                             | n/a                |
+| severity           | false    | Array of severity types for VULNS and ISSUES (ALL) import. Allowed values: CRITICAL,HIGH,MEDIUM,LOW,INFO. Import all if not present.                                                                                                                       | n/a                |
+| issue_status       | false    | Array of issue status for ISSUES import. Allowed values: OPEN,IN_PROGRESS,RESOLVED,REJECTED. Import all if not present.                                                                                                                                    | n/a                |
+| import_type        | false    | What to import, ISSUES, VULNS or ALL                                                                                                                                                                                                                       | ALL                |
+| external_id_attr   | false    | For ISSUES, the `entitySnapshot` attribute used to map Kenna asset's external_id, for instance, `providerId` or `resourceGroupExternalId`. If not present or the value for the passed attribute is not present the `providerId` attribute value is used.   | n/a                |
+| kenna_api_key      | false    | Kenna API Key for use with connector option                                                                                                                                                                                                                | n/a                |
+| kenna_api_host     | false    | Kenna API Hostname if not US shared                                                                                                                                                                                                                        | api.kennasecurity.com |
+| kenna_connector_id | false    | If set, we'll try to upload to this connector                                                                                                                                                                                                              | n/a                |
+| output_directory   | false    | If set, will write a file upon completion. Path is relative to #{$basedir}                                                                                                                                                                                 | output/wiz         |
