+++ tasks/connectors/wiz/wiz_task.rb
@@ -1,45 +1,72 @@
 # frozen_string_literal: true
 
 require_relative "lib/wiz_client"
+require_relative "lib/issues_mapper"
+require_relative "lib/vulns_mapper"
 module Kenna
   module Toolkit
     class WizTask < Kenna::Toolkit::BaseTask
       def self.metadata
         {
           id: "wiz",
-          name: "Wiz",
-          description: "Pulls assets and vulnerabilitiies from Wiz",
+          name: "WIZ",
+          description: "Pulls assets, vulnerabilities and issues from WIZ",
           options: [
             { name: "wiz_client_id",
               type: "string",
               required: true,
               default: nil,
-              description: "Wiz client id" },
+              description: "WIZ client id" },
             { name: "wiz_client_secret",
               type: "api_key",
               required: true,
               default: nil,
-              description: "Wiz client secret" },
+              description: "WIZ client secret" },
             { name: "wiz_auth_endpoint",
               type: "hostname",
               required: false,
               default: "auth.wiz.io",
-              description: "url to retrieve hosts and vulns - if no variation this might not need to be a param" },
+              description: "WIZ auth endpoint hostname used to get the authorization token." },
             { name: "wiz_api_host",
               type: "hostname",
               required: true,
-              default: "",
-              description: "url to retrieve hosts and vulns - find it here https://app.wiz.io/user/profile - API Endpoint URL" },
-            { name: "vulnerabilities_since",
+              default: nil,
+              description: "WIZ API Endpoint URL. If schema is included, it should be between double quotes escaped." },
+            { name: "wiz_page_size",
+              type: "integer",
+              required: false,
+              default: 500,
+              description: "Maximum number of issues to retrieve in each page." },
+            { name: "days_back",
               type: "integer",
               required: false,
               default: nil,
-              description: "integer days number to get the vulnerabilities detected SINCE x days" },
-            { name: "report_object_types",
+              description: "Integer days number to get the vulnerabilities/issues detected x days back TODAY." },
+            { name: "vuln_object_types",
               type: "string",
               required: false,
-              default: "VIRTUAL_MACHINE,CONTAINER_IMAGE,SERVERLESS",
-              description: "array of object types to include in the report" },
+              default: nil,
+              description: "Array of object types for VULNS import. Allowed values: VIRTUAL_MACHINE,CONTAINER_IMAGE,SERVERLESS. Import all if not present." },
+            { name: "severity",
+              type: "string",
+              required: false,
+              default: nil,
+              description: "Array of severity types for VULNS and ISSUES (ALL) import. Allowed values: CRITICAL,HIGH,MEDIUM,LOW,INFO. Import all if not present." },
+            { name: "issue_status",
+              type: "string",
+              required: false,
+              default: nil,
+              description: "Array of issue status for ISSUES import. Allowed values: OPEN,IN_PROGRESS,RESOLVED,REJECTED. Import all if not present." },
+            { name: "import_type",
+              type: "string",
+              required: false,
+              default: "ALL",
+              description: "What to import, ISSUES, VULNS or ALL" },
+            { name: "external_id_attr",
+              type: "string",
+              required: false,
+              default: nil,
+              description: "For ISSUES, the entitySnapshot attribute used to map Kenna asset's external_id, for instance, `providerId` or `resourceGroupExternalId`. If not present or the value for the passed attribute is not present the provideId attribute value is used." },
             { name: "kenna_api_key",
               type: "api_key",
               required: false,
@@ -64,141 +91,86 @@ module Kenna
         }
       end
 
-      def delete_file(dir, fname)
-        File.delete("#{dir}/#{fname}")
+      def run(opts)
+        super
+        initialize_options
+        initialize_client
+
+        import_issues if import?("issues")
+        import_vulns if import?("vulns")
+
+        kdi_connector_kickoff(@kenna_connector_id, @kenna_api_host, @kenna_api_key)
+      rescue Kenna::Toolkit::Sample::Client::ApiError => e
+        fail_task e.message
       end
 
-      def run(opts)
-        super # opts -> @options
-        # in this section get the options into variables if needed
-        # if you will only reference the data from this method you can call @options in-line
+      private
 
-        client_id = @options[:wiz_client_id]
-        client_secret = @options[:wiz_client_secret]
-        report_object_types = @options[:report_object_types].split(",")
+      attr_reader :client
+
+      def initialize_client
+        @client = Wiz::Client.new(@client_id, @client_secret, @auth_endpoint, @api_host, @page_size, @days_back, @vuln_object_types, @severity, @issue_status)
+      end
+
+      def initialize_options
+        @client_id = @options[:wiz_client_id]
+        @client_secret = @options[:wiz_client_secret]
+        @auth_endpoint = @options[:wiz_auth_endpoint].start_with?("http") ? "#{@options[:wiz_auth_endpoint]}/oauth/token" : "https://#{@options[:wiz_auth_endpoint]}/oauth/token"
+        @api_host = @options[:wiz_api_host].start_with?("http") ? "#{@options[:wiz_api_host]}/graphql" : "https://#{@options[:wiz_api_host]}/graphql"
+        @vuln_object_types = extract_list(:vuln_object_types)
+        @severity = extract_list(:severity)
+        @issue_status = extract_list(:issue_status)
+        @days_back = @options[:days_back].to_i if @options[:days_back].present?
+        @page_size = @options[:wiz_page_size].to_i if @options[:wiz_page_size].present?
+        @import_type = @options[:import_type].downcase
+        @external_id_attr = @options[:external_id_attr]
         @output_directory = @options[:output_directory]
         @kenna_api_host = @options[:kenna_api_host]
         @kenna_api_key = @options[:kenna_api_key]
         @kenna_connector_id = @options[:kenna_connector_id]
-        days_used_regenerate_report = false
-        vulnerabilities_since = ""
-        unless @options[:vulnerabilities_since].nil?
-          days_used_regenerate_report = true
-          now = Date.today
-          days_ago = (now - @options[:vulnerabilities_since].to_i)
-          vulnerabilities_since = days_ago.strftime("%FT%TZ")
-        end
-        skip_autoclose = false
-        retries = 3
-        kdi_version = 2
+        @batch_size = @options[:batch_size].to_i
+        @skip_autoclose = false
+        @retries = 3
+        @kdi_version = 2
+      end
 
-        client = Kenna::Toolkit::Wiz::WizClient.new(client_id, client_secret, @output_directory, @options[:wiz_auth_endpoint], @options[:wiz_api_host])
-        fail_task "Unable to retrieve API token, please check credentials" unless client.valid_token?
-        print_debug "report object types count #{report_object_types.size} and #{report_object_types}"
+      def extract_list(key, default = nil)
+        list = (@options[key] || "").split(",").map(&:strip)
+        list.empty? ? default : list
+      end
 
-        # @create_report_variables[:input][:params][:vulnerabilities_since] = vulnerabilities_since if vulnerabilities_since != ""
+      def import?(type)
+        @import_type == type || @import_type == "all"
+      end
 
-        client.create_report(days_used_regenerate_report, report_object_types, vulnerabilities_since)
+      def import_issues
+        print_good "Issues import started."
+        import(client.paged_issues, Wiz::IssuesMapper.new(@external_id_attr))
+      end
 
-        Dir.entries(@output_directory.to_s).each do |abspath|
-          next unless abspath.end_with? ".csv"
+      def import_vulns
+        print_good "Vulns import started."
+        import(client.paged_vulns, Wiz::VulnsMapper.new)
+      end
 
-          fname = File.basename(abspath)
-          csv_file = CSV.parse(File.open("#{@output_directory}/#{fname}", "r:bom|utf-8", &:read), headers: true)
-          unless csv_file.size.positive?
-            delete_file(@output_directory, fname)
-            next
-          end
-          csv_file.each do |row|
-            vuln_url = row["WizURL"]
-            cve = row["Name"]
-            severity = row["VendorSeverity"]
-            version = row["Version"]
-            fixed_version = row["FixedVersion"]
-            first_seen = row["FirstDetected"]
-            last_seen = row["LastDetected"]
-            solution = row["Remediation"]
-            hostname = row["AssetName"]
-            unique_id = row["ProviderUniqueId"]
-            image_id = ""
-            os = nil
-            runtime = ""
-            tags = []
-            tags_hash = JSON.parse(row["Tags"])
-            unless tags_hash.empty?
-              tags_hash.each do |key, value|
-                tags << "#{key}:#{value}"
-              end
-            end
-            tags << "Region:#{row['AssetRegion']}"
-            tags << "CloudPlatform:#{row['CloudPlatform']}"
-            vuln_severity = { "Critical" => 10, "High" => 8, "Medium" => 6, "Low" => 3 }
-            vuln_score = vuln_severity[severity].to_i
-            if abspath.include? "VIRTUAL_MACHINE"
-              runtime = row["Runtime"]
-              os = row["OperatingSystem"] unless row["OperatingSystem"].empty?
-            elsif abspath.include? "CONTAINER_IMAGE"
-              image_id = row["ImageId"]
-            elsif abspath.include? "SERVERLESS"
-              os = row["OperatingSystem"] unless row["OperatingSystem"].empty?
-              external_id = unique_id
-            end
-            asset = {
-              # used for VM assets primarily
-              "image_id" => image_id,
-              "hostname" => hostname,
-              "tags" => tags,
-              "external_id" => external_id
-            }
-            asset["os"] = row["OperatingSystem"] unless os.nil?
-            asset.compact!
-            details_additional_fields = {
-              "WizURL" => vuln_url,
-              "Version" => version,
-              "FixedVersion" => fixed_version,
-              "Projects" => row["Projects"],
-              "Runtime" => runtime,
-              "ProviderUniqueId" => unique_id,
-              "CloudProviderURL" => row["CloudProviderURL"]
-            }
-            # in case any values are null, it's good to remove them
-            details_additional_fields.compact!
+      def import(pages, mapper)
+        pos = 0
+        pages.each do |page|
+          total_count = page["totalCount"]
+          nodes = page["nodes"]
+          nodes.each do |node|
+            asset = mapper.extract_asset(node)
+            vuln = mapper.extract_vuln(node)
+            definition = mapper.extract_definition(node)
 
-            vuln = {
-              "scanner_type" => "Wiz",
-              "scanner_identifier" => cve,
-              # next is only needed for KDI V2 = vuln short name, text name, or cve or cwe name
-              "vuln_def_name" => cve,
-              "created_at" => first_seen,
-              "scanner_score" => vuln_score,
-              "last_seen_at" => last_seen,
-              "details" => JSON.pretty_generate(details_additional_fields)
-            }
-            # in case any values are null, it's good to remove them
-            vuln.compact!
-
-            vuln_def = {
-              # PICK (CVE OR CWE OR WASC) OR none but not all three
-              "cve_identifiers" => cve,
-              "solution" => solution,
-              "scanner_type" => "Wiz",
-              "name" => cve
-            }
-            # in case any values are null, it's good to remove them
-            vuln_def.compact!
-
-            # Create the KDI entries for vulns or findings
             create_kdi_asset_vuln(asset, vuln)
-
-            # create the KDI vuln def entry
-            create_kdi_vuln_def(vuln_def)
+            create_kdi_vuln_def(definition)
           end
-          filename = abspath.sub(/.csv/, ".json")
-          kdi_upload @output_directory, filename, @kenna_connector_id, @kenna_api_host, @kenna_api_key, skip_autoclose, retries, kdi_version
-          delete_file(@output_directory, fname)
+
+          print_good("Processed #{[pos + @page_size, total_count].min} of #{total_count} #{mapper.plural_name.downcase}.")
+          kdi_upload(@output_directory, "wiz_#{mapper.plural_name.downcase}_#{pos}.json", @kenna_connector_id, @kenna_api_host, @kenna_api_key, @skip_autoclose, @retries, @kdi_version)
+          pos += @page_size
         end
-        # this method will automatically use the stored array of uploaded files when calling the connector
-        kdi_connector_kickoff(@kenna_connector_id, @kenna_api_host, @kenna_api_key)
       end
     end
   end
